Remove dependence on Jersey CDI modules
CDI needs to know all known beans when it initializes. Unfortunately,
this is also true for Jersey. This creates a classic chicken-or-egg
problem where we can't start CDI before we start Jersey - meaning we have
to come up with a way to register all the injectable classes required by
Jersey BEFORE we initialize our own CDI context.

The solution here is to implement the JerseyInjectionManager from scratch
to identify all the classes needed by Jersey - then we can move those over
to our CDI context.