Implement the first couple of supplier bindings from Jersey
As far as I can gather, these function roughly like producers in CDI.
Unfortunately, the bean manager in CDI has no way of adding a producer
in that sense. However, we should be able to use the produceWith() method
in AfterBeanDiscover.addBean() with an appropriate callback to get the
same behavior.