Make sure the REST server only binds to one interface
The server was previously binding to 0.0.0.0 (all interfaces) which is
not very secure and forces the user to have a firewall instead to limit
access to the REST endpoints.

The createServer() in Jersey actually has no way of controlling this,
which forced us to recreate some of its functionality and use reflection
to access internal classes in Jersey.