Run multiple threads during test execution
We use the build-helper-maven-plugin to determine the number of cores
on the host system. We then use 30% of these threads to fire up test
classes. We only use a fraction because Netty as a framework already
employs heavy threading. Thus, firing up an excessive amount of test
threads is completely unnecessary.