Complete refactor and rewrite and move to org.unigrid.nativecdi
The previous method no longer works with either new versions of CDI or
GraalVM. The prefered approach to get everything to work is probably to
use a GraalVM feature (@AutomaticFeature).

This implements a first prototype. This still needs a lot of work, but
the general idea is to introduce a "weld-proxies.json" that specifices how
CDI proxy objects are constructed.

This is done in a somewhat similar way in Oracles Helidon to get CDI
to "work".