While still wrapped wrong, GraalVM now properly finds our beans
This is achieved by implementing a specialized ProxyService for native
Graal images. In the next step, this now needs to be combined with a
Graal "Feature" or a Arthur extension where we register these bean
classes and their wrapped proxy objects. For this simple test to run we
need to wrap the following:

Application => ProxyObject(Application) / WeldClientProxy(Application)
Event => WeldProxy(Event)
Instance => WeldProxy(Instance)