Further testing with CDI BCE's, proxies, substitutions and Graal features
Saving all the test code up to this point. There is a lot of commented
out code and a lot of different attempts to do things.

Did an attempt with using Arc (CDI Lite implementation). This does work.
However, I have a feeling it will blow up in our face when we try to run
it together with Jersey.

At the end of the day, native execution is making me pretty nervous and
considering all the issues it brings, maybe it's better to, for now,
solve this in an alternative way.

This is solvable with a little more time, but at the moment, I think we
need to move on.