Add JLink step (via Tentackle) to the native image module
This creates a zip file that we can include in the distribution that
includes an executable version of the application together with a run
script that is appropriate for the operating system.

As this also includes the Java runtime, this is the only thing we have
to include in our native distribution.