Refactor and introduce "common" project for files shared between modules
This avoids the duplication of code and logic. For now, we move the
ApplicationDirectory and Version classes to this module. This are shared
between the native image creator and the main Hedgehog project.

Introduces a ApplicationDirectoryProducer in the main Hedgehog module that
wraps ApplicationDirectory and creates a model in application scope.