Use ApplicationDirectory in S3 object and bucket services
Much cleaner than hard coding the path to $ENV(user.home)/s3data,
allowing us to mock the directory during testing.

While this is not how the data will be stored once it ends up on the
network, this is still a good temporary way to allow for testing of the
S3 endpoints.