Handle changing connections (IP address) on the server side
This is done via an implemented connection handler that runs as a parent
under the stream handler and pipeline.

For now, we don't do the client side at all. While we might look into
taking care of that as well, it will generally resolve itself and nodes
that fall away from the network are purged when a timeout is detected.