Introduce connection container
Rather than having P2PClient directly implementing the Connection type,
this now happens via the ConnectionContainer class. This class holds all
the logic to "access" and communicate with the connection.

This allows us to set up a ConnectionContainer for an already present
connection and place it in a node of the topology.